requirements.txt
flask>=3.0
flask-cors>=4.0
opencv-python-headless>=4.9
numpy>=1.26
Pillow>=10.0
reportlab>=4.1
pymupdf>=1.23
gunicorn>=21.0
